requirements.txt
flask
opencv-python-headless
pillow
numpy
tensorflow-cpu
scikit-learn
imagehash
matplotlib
gunicorn
